+++ src/babiesCheckValid.ts
@@ -3,7 +3,11 @@ import {
   ItemType,
   TrinketType,
 } from "isaac-typescript-definitions";
-import { getCollectibleItemType, log } from "isaacscript-common";
+import {
+  getCollectibleItemType,
+  getCollectibleName,
+  log,
+} from "isaacscript-common";
 import { BABIES } from "./objects/babies";
 import { BabyDescription } from "./types/BabyDescription";
 
@@ -123,8 +127,18 @@ function logBabyInvalid(baby: BabyDescription, i: string, msg: string) {
 
 function logSpecificBabies() {
   for (const [i, baby] of Object.entries(BABIES)) {
-    if ("trinket" in baby && !("class" in baby) && !("item" in baby)) {
-      log(`DEBUG: ${baby.name} (#${i}) - ${TrinketType[baby.trinket]}`);
+    if (
+      "item" in baby &&
+      !("class" in baby) &&
+      !("trinket" in baby) &&
+      !("item2" in baby) &&
+      !("itemNum" in baby)
+    ) {
+      let collectibleName = getCollectibleName(baby.item);
+      if (collectibleName === "Unknown") {
+        collectibleName = `#${baby.item}`;
+      }
+      log(`DEBUG: ${baby.name} (#${i}) - ${collectibleName}`);
     }
   }
 }